more code cleanup, added logo to home screen, began layout column
