--- gui/src/App.js
@@ -1,68 +1,0 @@
-import React, { useState, useEffect } from 'react';
-import { Switch, Route } from 'react-router-dom';
-import './style.css';
-
-import { Header } from './components/Header.js';
-import { Home } from './pages/Home.js';
-import { Recipes } from './pages/Recipes.js';
-import { RecipeDetails } from './pages/RecipeDetails.js';
-import { ReactionHistory } from './pages/ReactionHistory'
-import { Tests } from './Tests.js';
-import { Settings } from './pages/Settings.js';
-import { Status } from './Status.js';
-
-import { apiUrl } from './utils.js';
-
-import './style.css';
-
-export function App() {
-	const [status, setStatus] = useState();
-	const getStatus = () => {
-		fetch(apiUrl + 'status')
-			.then(response => response.json())
-			.then(data => setStatus(data));
-	};
-	useEffect(() => {
-		const interval = setInterval(() => {
-			getStatus();
-		}, 1000);
-		return () => clearInterval(interval);
-	}, []);
-	console.log("status: ", status);
-	//these are the routes. is it in a div instead of Router for any reason?
-	//^i think becuase in style.css it sets the maybe-size of the screen
-	return (
-		<div className='lcd-wrapper'>
-			<Header>Lab status: {status?.status || 'loading...'}</Header> 
-			<Switch>
-				<Route exact path='/'>
-					<Home status={status} />
-				</Route>
-
-				<Route exact path='/recipes'>
-					<Recipes />
-				</Route>
-
-				<Route path='/recipes/:recipeName'>
-					<RecipeDetails />
-				</Route>
-
-				<Route path='/reaction-history'>
-					<ReactionHistory />
-				</Route>
-
-				<Route path='/tests'>
-					<Tests />
-				</Route>
-
-				<Route path='/settings'>
-					<Settings />
-				</Route>
-
-				<Route path='/status'>
-					<Status status={status}/>
-				</Route>
-			</Switch>
-		</div>
-	);
-}
