+++ gui/src/Stirring.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { Switch, Route, Link } from 'react-router-dom';
 import { Button, Icon, Menu, Segment } from 'semantic-ui-react';
 import './style.css';
-import { Header } from './Header.js';
+import { Header } from './Header';
 
 export function Settings() {
 	return (
