+++ gui/src/Therm.jsx
@@ -1,0 +1,15 @@
+import React from "react";
+import { Switch, Route, Link } from "react-router-dom";
+import { Button, Icon, Menu, Segment } from "semantic-ui-react";
+import "./style.css";
+import { Header } from "./Header";
+
+export function Settings() {
+  return (
+    <div>
+      <Header>Therm</Header>
+
+      <p>Therm</p>
+    </div>
+  );
+}
