+++ gui/src/components/Column.jsx
@@ -1,0 +1,8 @@
+import * as React from "react";
+import "../styles/column.css";
+
+const Column = ({ span, children }) => (
+  <div className={`column column-span-${span}`}>{children}</div>
+);
+
+export default Column;
