--- gui/src/index.js
@@ -1,6 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import { App } from './App.js';
-import { BrowserRouter as Router } from 'react-router-dom';
-
-ReactDOM.render(<Router><App /></Router>, document.getElementById('root'));
