+++ gui/src/pages/RecipeDetails.jsx
@@ -3,7 +3,7 @@ import { useParams} from 'react-router-dom';
 import { useHistory } from 'react-router-dom';
 import { Button } from 'semantic-ui-react';
 
-import { apiUrl } from '../utils.js';
+import { apiUrl } from '../utils';
 
 export function RecipeDetails() {
     const [recipeDetails, setRecipeDetails] = useState(false);
